[MCP-94] Refactor monolithic server into modular architecture
Transform single 1101-line server.py into maintainable modular structure:

* Create mcp_bluesky/ package with 13 focused modules
* Separate core functionality: config, client, types, context
* Organize tools by category: auth, profiles, posts, interactions, feeds, media
* Maintain 100% backward compatibility with existing server.py entry point
* Update all 39 tests with proper module mocking and import fixes
* Enhance documentation with new architecture details
* Add comprehensive refactoring plan documentation

Key improvements:
- Better separation of concerns and maintainability
- Easier testing with modular imports and dependency injection
- Clear tool categorization and navigation
- Enhanced developer experience with smaller, focused files
- Foundation for future enhancements and team collaboration

All tests pass, linting clean, full backward compatibility preserved.
