+++ mcp_bluesky/__init__.py
@@ -1,0 +1,6 @@
+"""Bluesky MCP Server Package.
+
+A modular MCP server implementation for interacting with the Bluesky social network.
+"""
+
+__version__ = "0.1.0"
