+++ mcp_bluesky/tools/__init__.py
@@ -1,0 +1,4 @@
+"""MCP Tools for Bluesky operations.
+
+This module contains all the tool implementations organized by functionality.
+"""
