+++ pyproject.toml
@@ -26,6 +26,7 @@ requires = ["setuptools>=45", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
+packages = ["mcp_bluesky"]
 py-modules = ["server"]
 
 [project.scripts]
