+++ tests/conftest.py
@@ -229,7 +229,7 @@ class MockBlueskyClient:
         feed_data = [
             {
                 "post": {
-                    "uri": f"at://did:plc:test123456789/app.bsky.feed.post/author1",
+                    "uri": "at://did:plc:test123456789/app.bsky.feed.post/author1",
                     "cid": "bafyauthor123",
                     "author": {
                         "did": "did:plc:test123456789",
@@ -423,5 +423,11 @@ def mock_bluesky_client():
 @pytest.fixture
 def mock_auth_client(mock_bluesky_client):
     """Fixture that mocks the get_authenticated_client function."""
-    with patch('server.get_authenticated_client', return_value=mock_bluesky_client):
+    # Patch the function where it's imported and used
+    with patch('mcp_bluesky.tools.auth.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.profiles.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.posts.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.interactions.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.feeds.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.media.get_authenticated_client', return_value=mock_bluesky_client):
         yield mock_bluesky_client
