+++ tests/test_feed_operations.py
@@ -1,7 +1,6 @@
 """Tests for feed operations: get_author_feed, get_post_thread."""
 
 import json
-import os
 import pytest
 import asyncio
 from unittest.mock import Mock, patch
@@ -27,7 +26,7 @@ class MockBlueskyClient:
         feed_data = [
             {
                 "post": {
-                    "uri": f"at://did:plc:test123456789/app.bsky.feed.post/author1",
+                    "uri": "at://did:plc:test123456789/app.bsky.feed.post/author1",
                     "cid": "bafyauthor123",
                     "author": {
                         "did": "did:plc:test123456789",
@@ -107,7 +106,13 @@ def mock_bluesky_client():
 @pytest.fixture
 def mock_auth_client(mock_bluesky_client):
     """Fixture that mocks the get_authenticated_client function."""
-    with patch('server.get_authenticated_client', return_value=mock_bluesky_client):
+    # Patch the function where it's imported and used
+    with patch('mcp_bluesky.tools.auth.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.profiles.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.posts.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.interactions.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.feeds.get_authenticated_client', return_value=mock_bluesky_client), \
+         patch('mcp_bluesky.tools.media.get_authenticated_client', return_value=mock_bluesky_client):
         yield mock_bluesky_client
